refactor(auth): consolidate stripe files into single directory
