+++ packages/auth/src/stripe/client.ts
@@ -1,0 +1,5 @@
+import Stripe from "stripe";
+
+const secretKey = process.env.STRIPE_SECRET_KEY || "sk_test_dummykey1234567890";
+
+export const stripeClient = new Stripe(secretKey, { apiVersion: "2026-01-28.clover" });
